:sparkles: feat : 어카운트 컨트롤러 작성 끝

+++ backend/src/main/java/org/example/b104/domain/account/service/command/AccountTransferCommand.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class AccountTransferCommand {
     String depositBankCode;
     String depositAccountNo;
-    int transactionBalance;
+    long transactionBalance;
     String withdrawalBankCode;
     String withdrawalAccountNo;
     String depositTransactionSummary;
